# docs/openapi/respondent.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "qcms respondent API",
    "version": "0.0.0",
    "description": "The public respondent loop: start a session, walk the branching flow, submit. Plus the always-on liveness/readiness probes."
  },
...
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "unauthorized"
              },
              "message": {
                "type": "string",
                "example": "Missing or invalid internal service token"
              },
              "details": {
                "nullable": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "StepResponse": {
        "type": "object",
        "properties": {
          "step": {
            "$ref": "#/components/schemas/StepDocument"
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "pattern": "^opt_[a-z0-9_]+$"
                  }
                }
              ]
            },
            "description": "The answers the server currently holds for this step's visible questions, keyed by questionId, in canonical AnswerValue encoding (an NFC string for shortText, longText, date and singleChoice; a finite number; a boolean; a duplicate-free array of optionIds for multiChoice). Absent keys are unanswered (including retracted answers), and a value is never null. Display data only; the flow projection stays the sole authority on visibility and readiness.",
            "example": {
              "q_at_fault_accident": true,
              "q_accident_count": 2
            }
          },
          "a2uiSpecVersion": {
            "type": "string",
            "description": "The pinned snapshot's A2UI spec version, so the renderer selects the right handling (ADR-18).",
            "example": "1.0.0-preview.7"
          },
          "flowState": {
            "$ref": "#/components/schemas/FlowStateProjection"
          },
          "rosters": {
            "$ref": "#/components/schemas/RosterProjection"
          },
          "progress": {
            "$ref": "#/components/schemas/StepProgress"
          }
        },
        "required": [
          "step",
          "values",
          "a2uiSpecVersion",
          "flowState",
          "rosters",
          "progress"
        ]
      },
      "StepDocument": {
        "type": "object",
        "nullable": true,
        "properties": {
          "stepId": {
            "type": "string",
            "example": "stp_history"
          },
          "root": {
            "nullable": true
          }
        },
        "required": [
          "stepId"
        ]
      },
      "FlowStateProjection": {
        "type": "object",
        "properties": {
          "currentStep": {
            "type": "string",
            "nullable": true,
            "example": "stp_history"
          },
          "visibleQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The currently-visible fields of the rendered step, as ANSWER KEYS: a bare questionId outside every repeating group, and `instanceId/questionId` inside one (ADR-42, ADR-43). A rule targeting inside a group is evaluated once per live instance, so a member question can be visible in one instance and hidden in another and only the qualified key can say which.",
            "example": [
              "q_at_fault_accident",
              "q_accident_count"
            ]
          },
          "missingRequired": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The visible required fields still unanswered, as answer keys, on the same rule as visibleQuestions: one entry per unanswered (instance, question) inside a group rather than one per question.",
            "example": [
              "q_accident_count"
            ]
          },
          "readyToSubmit": {
            "type": "boolean",
            "description": "True when no visible required question is unanswered (the flow may be submitted)."
          }
        },
        "required": [
          "currentStep",
          "visibleQuestions",
          "missingRequired",
          "readyToSubmit"
        ]
      },
      "RosterProjection": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "groupId": {
              "type": "string",
              "example": "grp_passengers"
            },
            "instances": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": [
                "ins_7k2",
                "ins_9m4"
              ]
            }
          },
          "required": [
            "groupId",
            "instances"
          ]
        }
      },
      "StepProgress": {
        "type": "object",
        "properties": {
          "stepIndex": {
            "type": "integer",
            "description": "0-based index of the current step within the visible steps; equals totalVisibleSteps when complete."
          },
          "totalVisibleSteps": {
            "type": "integer",
            "example": 1
          }
        },
        "required": [
          "stepIndex",
          "totalVisibleSteps"
        ]
      },
      "SubmitAnswerBody": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "minLength": 1,
            "example": "q_at_fault_accident"
          },
          "instanceId": {
            "type": "string",
            "minLength": 1,
            "description": "Which instance of a repeating group this answer belongs to (ADR-42). Absent for a question outside every group, which keeps the body byte-identical for every form that has none. A retraction names an instance too, and clears that cell alone (ADR-33's Note).",
            "example": "ins_7k2"
          },
          "value": {
            "nullable": true,
            "description": "The answer value; validated against the pinned question. Literal null retracts the answer (the question becomes unanswered; the ledger records the retraction). An empty value (\"\" or []) is not an answer and is rejected with EMPTY_ANSWER_NOT_ALLOWED; send null to clear an answer."
          }
        },
        "required": [
          "questionId"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/sessions/{id}/answers": {
      "post": {
        "summary": "Submit one answer (or retract it with a null value); validated, appended to the ledger, flow re-evaluated (session-token authed)",
        "tags": [
          "responses"
        ],
        "security": [
          {
            "MachineToken": [
              "responses:write"
            ]
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "ses_9f3a2b1c"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "description": "0-based index of the visible step to render (the explicit navigation cursor, ADR-28). Clamped to the visible range; omit to serve the first incomplete step.",
              "example": 1
            },
            "required": false,
            "description": "0-based index of the visible step to render (the explicit navigation cursor, ADR-28). Clamped to the visible range; omit to serve the first incomplete step.",
            "name": "step",
            "in": "query"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The answer was recorded; the updated flow projection follows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}